Déplace réductions vers son propre fichier
